api-gateway/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
pydantic==1.8.2
httpx==0.19.0
python-jose==3.3.0
python-multipart==0.0.5
python-dotenv==0.19.0